refactor RDSeffSwitch
