+++ bundles/org.palladiosimulator.simulizar/src/org/palladiosimulator/simulizar/interpreter/GetParentSwitchInterface.java
@@ -1,0 +1,8 @@
+package org.palladiosimulator.simulizar.interpreter;
+
+
+import org.eclipse.emf.ecore.util.Switch;
+
+public interface GetParentSwitchInterface {
+  public Switch<Object> getParentSwitch();
+}
